wijzig gerechten en ingredienten

+++ backend/src/services/gerecht/gerecht.service.ts
@@ -3,7 +3,7 @@ import { Gerecht } from 'src/models/gerecht';
 import { CreateGerechtDto } from 'src/dto/create-gerecht-dto';
 import { GerechtRepoEntity } from 'src/entities/gerecht.entity';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository, DeleteResult } from 'typeorm';
+import { Repository, DeleteResult, UpdateResult, Not } from 'typeorm';
 
 @Injectable()
 export class GerechtService {
@@ -28,7 +28,20 @@ export class GerechtService {
         }
     }
 
-    deleteGerecht(id: number): Promise<DeleteResult> {
-        return this.gerechtRepository.delete({id});
+    async updateGerecht(id: number, gerechtEntity: GerechtRepoEntity): Promise<UpdateResult> {
+        const gerechtExists = !!(await this.gerechtRepository.findOne({ id }));
+        if (!gerechtExists) throw new HttpException('Er bestaat geen gerecht met dit id', HttpStatus.NOT_FOUND);
+
+        const naamExists = !!(await this.gerechtRepository.findOne({ id: Not(id), naam: gerechtEntity.naam }));
+        if (naamExists) throw new HttpException('Een gerecht met deze naam bestaat al', HttpStatus.CONFLICT);
+
+        return this.gerechtRepository.update({ id }, gerechtEntity);
+    }
+
+    async deleteGerecht(id: number): Promise<DeleteResult> {
+        const gerechtExists = !!(await this.gerechtRepository.findOne({ id }));
+        if (!gerechtExists) throw new HttpException('Er bestaat geen gerecht met dit id', HttpStatus.NOT_FOUND);
+
+        return this.gerechtRepository.delete({ id });
     }
 }
